refactor: move and rename adapters and routes
